[lookup] move lookup index to its own module
