+++ kimchi/src/circuits/lookup/mod.rs
@@ -2,5 +2,6 @@
 //! See <https://eprint.iacr.org/2020/315.pdf>
 
 pub mod constraints;
+pub mod index;
 pub mod lookups;
 pub mod tables;
